# [shared] PropertyFetcher perf tweak