## Changes

* Tweaks `PropertyFetcher` to have a little bit better perf

## Benchmarks

net462 build via .NET Framework 4.8.1 runtime:

|                      Method |     Mean |    Error |   StdDev | Allocated |
|---------------------------- |---------:|---------:|---------:|----------:|
|        UsingPropertyFetcher | 14.72 ns | 0.173 ns | 0.144 ns |         - |
| UsingTweakedPropertyFetcher | 11.66 ns | 0.278 ns | 0.260 ns |         - |

net6.0 build via .NET 7.0.400 sdk:

|                                     Method |      Mean |     Error |    StdDev | Allocated |
|------------------------------------------- |----------:|----------:|----------:|----------:|
|        UsingPropertyFetcher | 11.646 ns | 0.1534 ns | 0.1281 ns |         - |
|  UsingTweakedPropertyFetcher |  7.326 ns | 0.1401 ns | 0.1311 ns |         - |

<details>
<summary>Dragons</summary>

This work originally spawned out of https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1251#discussion_r1296266126 where I noticed a lower-level version of `PropertyFetcher` was _much_ faster.

There isn't a whole lot to the benchmarks, just running the same code using the existing version and the new one:

```csharp
[MemoryDiagnoser]
public class WcfReflectionBenchmarks
{
    private readonly PropertyFetcher<WebHeaderCollection> propertyFetcher = new("Headers");
    private readonly PropertyFetcherTweaks<WebHeaderCollection> propertyFetcherTweaks = new("Headers");
    private readonly HttpRequestMessageProperty request = new();

    [Benchmark]
    public WebHeaderCollection UsingPropertyFetcher()
    {
        if (!this.propertyFetcher.TryFetch(this.request, out var value))
        {
            throw new InvalidOperationException();
        }

        return value;
    }

    [Benchmark]
    public WebHeaderCollection UsingTweakedPropertyFetcher()
    {
        if (!this.propertyFetcherTweaks.TryFetch(this.request, out var value))
        {
            throw new InvalidOperationException();
        }

        return value;
    }
}
```
</details>

## Merge requirement checklist

* [X] [CONTRIBUTING](https://github.com/open-telemetry/opentelemetry-dotnet/blob/main/CONTRIBUTING.md) guidelines followed (nullable enabled, static analysis, etc.)
* [X] Unit test improvements