@utpilla 2023-08-17
> ## Changes
> * Tweaks `PropertyFetcher` to have a little bit better perf

Where's the perf improvement mainly coming from? Is it from using a variable on stack to hold the reference to `this.innerFetcher` vs always referring to the instance field?