@utpilla 2023-08-17
> ## Benchmarks
> Ran on net462 build via .NET Framework 4.8.1 runtime
> 
> Method	Mean	Error	StdDev	Allocated
> UsingPropertyFetcher	15.028 ns	0.2683 ns	0.4016 ns	-
> UsingTweakedPropertyFetcher	12.567 ns	0.2910 ns	0.3573 ns	-

Could you also add the benchmark code in the description?