@CodeBlanch 2023-08-18
> Where's the perf improvement mainly coming from? Is it from using a variable on stack to hold the reference to this.innerFetcher vs always referring to the instance field?

The local thing is minor if it changes anything at all. I actually just improved it to use a sub-method. The sub-method for the rare case allows the JIT to be more aggressive in its inline decision with the goal being it should inline the hot-path but not the rare case. It also reduced some code duplication which is nice!

Running on .NET Framework 4.8.1 it doesn't look like the JIT is smart enough. But on net6 I see it improving:

|                                     Method |      Mean |     Error |    StdDev | Allocated |
|------------------------------------------- |----------:|----------:|----------:|----------:|
|        UsingPropertyFetcher | 11.646 ns | 0.1534 ns | 0.1281 ns |         - |
|  UsingRarePropertyFetcher |  7.326 ns | 0.1401 ns | 0.1311 ns |         - |
| UsingTweakedPropertyFetcher |  8.428 ns | 0.1655 ns | 0.1548 ns |         - |

Where is the perf boost coming from?

Short answer is: Less operations in the hot path.

Long answer:

We can make some assumptions about how this is used:

* Payload object being supplied is not null.
* After the first call we should have an inner fetcher.

The version we have today isn't optimized for those assumptions. It always does a `null` check or checks a `bool` if it should skip that check. And it checks twice if it has a valid inner fetcher. The new version does 1 check to see if it has an inner fetcher and if it does it immediately calls it. The first operation done by the inner fetcher is `if (obj is TDeclaredObject o)` which will only be `true` when `obj` is not `null` so we don't really need an extra null check for the happy/hot path. We only need the `null` check in the rare case where we need to create an inner fetcher because we need the type info (`obj.GetType()`).